backend/requirements.txt
attrs==25.3.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
docx2txt==0.8
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.0.9
packaging==24.2
pandas==2.0.3
pdf2image==1.16.0
Pillow==8.3.2
pluggy==1.5.0
py==1.11.0
pytesseract==0.3.8
pytest==6.2.5
pytest-flask==1.2.0
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2025.2
requests==2.32.3
six==1.17.0
toml==0.10.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
